% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{fit_BCP_INGARCH}
\alias{fit_BCP_INGARCH}
\title{Fit a Bivariate Conditional Poisson INGARCH(1,1) model}
\usage{
fit_BCP_INGARCH(Y, A.diag = TRUE, B.diag = FALSE)
}
\arguments{
\item{Y}{integer matrix of bivariate count data.}

\item{A.diag}{boolean}

\item{B.diag}{boolean}
}
\description{
Fit a Bivariate Conditional Poisson INGARCH(1,1) model
}
