% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_se.R
\name{get_boot_se}
\alias{get_boot_se}
\title{Bootstrap standard errors of BCP-INGARCH(1,1) CMLEs}
\usage{
get_boot_se(model_fit, Y, replicas)
}
\arguments{
\item{model_fit}{Fitted BCP-INGARCH(1,1) model}

\item{Y}{observed bivariate count data}

\item{replicas}{Number of bootstrap replicas}
}
\description{
Bootstrap standard errors of BCP-INGARCH(1,1) CMLEs
}
