% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{model_information}
\alias{model_information}
\title{Model information criteria for BCP-INGARCH(1,1) model fits.}
\usage{
model_information(fit)
}
\arguments{
\item{fit}{fitted model obtained with fit_BCP_INGARCH.}
}
\description{
Model information criteria for BCP-INGARCH(1,1) model fits.
}
