% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcpingarch-package.R
\docType{package}
\name{rbcpingarch-package}
\alias{rbcpingarch-package}
\alias{rbcpingarch}
\title{The 'rbcpingarch' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2022). RStan: the R interface to Stan. R package version 2.21.5. https://mc-stan.org
}
