% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sim.R
\name{rBCPINGARCH}
\alias{rBCPINGARCH}
\title{Simulate a Bivariate Conditional Poisson INGARCH(1,1) process}
\usage{
rBCPINGARCH(A, B, omega, phi, n)
}
\arguments{
\item{A}{2x2 matrix of parameters}

\item{B}{2x2 matrix of parameters}
}
\description{
Simulate a Bivariate Conditional Poisson INGARCH(1,1) process
}
